Some modifications from OOP point of view
